services/world_cup_stats_raw_to_domain_service/requirements.txt
google-cloud-bigquery==3.5.0
google-cloud-storage==2.10.0
fastapi==0.95.2
uvicorn==0.22.0
dacite==1.6.0
toolz==0.12.0
ndjson==0.3.1
